move ShopifyError up from SDK
